iextrading4j-client/src/test/java/pl/zankowski/iextrading4j/client/sse/request/stocks/NewsSseRequestBuilderTest.java
package pl.zankowski.iextrading4j.client.sse.request.stocks;

import org.junit.Test;
import pl.zankowski.iextrading4j.api.stocks.v1.News;
import pl.zankowski.iextrading4j.client.sse.manager.SseRequest;

import javax.ws.rs.core.GenericType;

import static org.assertj.core.api.Assertions.assertThat;
import static org.assertj.core.api.Assertions.entry;

public class NewsSseRequestBuilderTest {

    @Test
    public void shouldSuccessfullyCreateRequest() {
        final String symbol = "IBM";

        final SseRequest<News> request = new NewsSseRequestBuilder()
                .withSymbol(symbol)
                .build();

        assertThat(request.getPath()).isEqualTo("/news-stream");
        assertThat(request.getResponseType()).isEqualTo(new GenericType<News>() {
        });
        assertThat(request.getPathParams()).isEmpty();
        assertThat(request.getQueryParams()).contains(entry("nosnapshot", "false"), entry("symbols", symbol));
    }

}
